+++ maxscale-system-test/mariadb_nodes.cpp
@@ -698,7 +698,6 @@ static bool wrong_replication_type(MYSQL *conn)
         }
         sleep(1);
     }
-
     return rval;
 }
 
@@ -1394,7 +1393,7 @@ int Mariadb_nodes::prepare_server(int i)
     char str1[1024];
     char str2[1024];
 
-    ssh_node(i, stop_db_command[i], true);
+    ssh_node_f(i, true, "%s", stop_db_command[i]);
     sleep(5);
     ssh_node(i, "sed -i \"s/bind-address/#bind-address/g\" /etc/mysql/my.cnf.d/*.cnf", true);
     ssh_node(i, "ln -s /etc/apparmor.d/usr.sbin.mysqld /etc/apparmor.d/disable/usr.sbin.mysqld; sudo service apparmor restart", true);