+++ server/core/service.cc
@@ -2654,13 +2654,13 @@ json_t* service_attributes(const SERVICE* service)
     json_object_set_new(attr, CN_ROUTER, json_string(service->routerModule));
     json_object_set_new(attr, CN_STATE, json_string(service_state_to_string(service->state)));
 
-    if (service->router && service->router_instance)
+    if (service->router && service->router_instance && service->router->diagnostics_json)
     {
         json_t* diag = service->router->diagnostics_json(service->router_instance);
 
         if (diag)
         {
-            json_object_set_new(attr, "router_diagnostics", diag);
+            json_object_set_new(attr, CN_ROUTER_DIAGNOSTICS, diag);
         }
     }